+++ CHANGELOG.rst
@@ -114,6 +114,9 @@ New features:
 - Improved performance of waiting for processes (on Windows by blocking with WaitForSingleObject instead of sleep polling, 
   and on Linux if kernel version is 5.3+ and Python=3.9 using pidfd_open).
 
+- Added project property attribute ``path=`` as an alternative to ``value=`` for properties containing a path, which 
+  should be normalized to remove any ``/../`` path components. 
+
 Fixes:
 
 - Fixed the GitHub Actions support to stop using the recently deprecated ``::set-output`` mechanism for publishing